requirements.txt
streamlit
altair
transformers
torch
numpy
faiss-cpu
sentence-transformers
